로그인/회원가입 프래그먼트에 api 연동
