+++ app/src/main/java/com/example/umc_hackathon/post/FormDetailActivity.kt
@@ -5,14 +5,10 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import android.view.View
-import androidx.core.view.isInvisible
-import androidx.core.view.isVisible
 import com.example.umc_hackathon.auth.AuthActivity
-import com.example.umc_hackathon.auth.LoginActivity
 import com.example.umc_hackathon.databinding.ActivityFormDetailBinding
 import com.example.umc_hackathon.post.*
 import com.example.umc_hackathon.post.result.ResultActivity
-import com.example.umc_hackathon.post.result.ResultFragment
 import com.example.umc_hackathon.survey.FormInputActivity
 
 class FormDetailActivity : AppCompatActivity(), PostDetailView {
@@ -133,7 +129,7 @@ class FormDetailActivity : AppCompatActivity(), PostDetailView {
                 binding.formDetailDislikeBtnCv.visibility = View.VISIBLE
             }
 
-            // status = closed일 때 button 비활성화
+            // status == closed일 때 button 비활성화
             if(result.status == "CLOSED") {
                 binding.formDetailParticipateBtn.isEnabled = false
             }
